backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
